Get average buffered urban imperviousness

The percent urban imperviousness of each NLCD 30 x 30 m cell within buffer meters of locations in x are averaged.

Usage

get_urban_imperviousness(x, year, buffer = 400)

install_urban_imperviousness(year = as.character(c(2021, 2019, 2016)))

Arguments

x

a vector of s2 cell identifiers (s2_cell object)

year

a character string that is the year of the urban imperviousness data

buffer

distance from s2 cell (in meters) to summarize data

Value

for get_urban_imperviousness(), a numeric vector the same length as x

for install_impervious(), a character string path to impervious raster data

Details

Urban imperviousness data is released sequentially, so the 2021 data is retrieved from the 2023-06-30 release; both the 2019 and 2016 data are retrieved from the 2021-06-04 release. From the metadata attribute definition for percent imperviousness: "while the file structure shows values in range from 0-255, the values of 0-100 are the only real populated values, in addition to a background value of 127". Extracted values greater than 100 are changed to 0 before averaging.

References

https://www.mrlc.gov/data/type/urban-imperviousness

https://www.mrlc.gov/viewer/

https://www.usgs.gov/centers/eros/science/national-land-cover-database

Examples

get_urban_imperviousness(
  x = s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")),
  year = "2021"
)
#> 8841b399ced97c47 8841b38578834123 
#>         74.32258         33.62770