Assemble a tibble of required predictors for the exposure assessment model

Assemble a tibble of required predictors for the exposure assessment model

Usage

assemble_predictors(x, dates, pollutant = c("pm25"), quiet = TRUE)

Arguments

x

a vector of s2 cell identifers (s2_cell object); currently required to be within the contiguous united states

dates

a list of date vectors for the predictions, must be the same length as x

pollutant

ignored now, but reserved for future sets of predictors specific to different pollutants

quiet

silence progress messages?

Value

a tibble with one row for each unique s2 location - date combination where columns are predictors required for the exposure assessment model

Examples

d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
  "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
)
assemble_predictors(x = s2::as_s2_cell(names(d)), dates = d) |>
  tibble::glimpse()
#> downloading 2023 air.2m:
#> downloading 2023 hpbl:
#> downloading 2023 acpcp:
#> downloading 2023 rhum.2m:
#> downloading 2023 vis:
#> downloading 2023 pres.sfc:
#> downloading 2023 uwnd.10m:
#> downloading 2023 vwnd.10m:
#> Rows: 4
#> Columns: 28
#> $ s2                       <s2cell> 8841b39a7c46e25f, 8841b39a7c46e25f, 8841a4…
#> $ date                     <date> 2023-05-18, 2023-11-06, 2023-06-22, 2023-08-…
#> $ x                        <dbl> 981664.4, 981664.4, 993445.3, 993445.3
#> $ y                        <dbl> 1850030, 1850030, 1833221, 1833221
#> $ elevation_median_800     <dbl> 231.0, 231.0, 231.5, 231.5
#> $ elevation_sd_800         <dbl> 11.16542, 11.16542, 12.88410, 12.88410
#> $ aadt_total_m_400         <dbl> 16106441, 16106441, 0, 0
#> $ aadt_truck_m_400         <dbl> 1.394319e+13, 1.394319e+13, 0.000000e+00, 0.0…
#> $ air.2m                   <dbl> 286.4203, 283.4198, 291.6172, 297.2624
#> $ hpbl                     <dbl> 654.3253, 536.9161, 775.3117, 1110.5875
#> $ acpcp                    <dbl> 0.002148881, 0.000000000, 0.000000000, 2.4609…
#> $ rhum.2m                  <dbl> 72.11295, 63.10049, 81.88810, 78.28357
#> $ vis                      <dbl> 20013.14, 20006.17, 20016.76, 19812.84
#> $ pres.sfc                 <dbl> 98413.07, 98460.16, 98972.92, 98474.33
#> $ uwnd.10m                 <dbl> -3.310530901, 0.002454758, -4.644292831, 5.37…
#> $ vwnd.10m                 <dbl> -1.10644674, 2.38253164, -1.65637243, -0.0577…
#> $ merra_dust               <dbl> 1.7716578, 0.8419505, 1.3394157, 1.1787193
#> $ merra_oc                 <dbl> 6.839098, 2.653597, 2.519055, 2.797475
#> $ merra_bc                 <dbl> 0.5318720, 0.3922745, 0.3267147, 0.4414604
#> $ merra_ss                 <dbl> 0.9944661, 0.2441934, 0.3561678, 0.7222975
#> $ merra_so4                <dbl> 2.430319, 2.208039, 3.706336, 5.782006
#> $ urban_imperviousness_400 <dbl> 66.63441, 66.63441, 11.82796, 11.82796
#> $ nei_point_id2w_1000      <dbl> 2.87343e-05, 2.87343e-05, 0.00000e+00, 0.0000…
#> $ smoke_pm                 <dbl> 0, 0, 0, 0
#> $ year                     <dbl> 2023, 2023, 2023, 2023
#> $ doy                      <dbl> 138, 310, 173, 227
#> $ month                    <dbl> 5, 11, 6, 8
#> $ dow                      <dbl> 4, 1, 4, 2