Get smoke plume data from the NOAA's Hazard Mapping System

The HMS operates daily in near real-time by outlining the smoke polygon of each distinct smoke plume and classifying it as "light", "medium", and "heavy". Since multiple plumes of varying or the same classification can cover one another, the total smoke plume exposure is estimated as the weighted sum of all plumes, where "light" = 1, "medium" = 2, and "heavy" = 3.

Usage

get_hms_smoke_data(x, dates, quiet = TRUE)

install_hms_smoke_data()

Arguments

x

a vector of s2 cell identifers (s2_cell object); currently required to be within the contiguous united states

dates

a list of date vectors for the predictions, must be the same length as x

quiet

silence progress messages?

Value

for get_hms_smoke_data(), a list of numeric vectors of smoke plume scores (the same length as x and dates)

for install_hms_smoke_data(), a character string path to the installed RDS file

Details

Daily HMS shapefiles are missing for 7 days within 2017-2023 ("2017-04-27", "2017-05-31", "2017-06-01", "2017-06-01" "2017-06-22", "2017-11-12", "2018-12-31") and will return zero values. If files are available but no smoke plumes intersect, then a zero values is also returned.

References

https://www.ospo.noaa.gov/Products/land/hms.html#about

Examples

d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2017-11-06")),
  "8841a45555555555" = as.Date(c("2017-06-22", "2023-08-15"))
)
get_hms_smoke_data(x = s2::as_s2_cell(names(d)), dates = d)
#> [[1]]
#> [1] 9 0
#> 
#> [[2]]
#> [1] 0 2
#>