Get traffic summary

Highway Performance Monitoring System (HPMS) data from 2020 is summarized as the sum of the total (and truck-only) average annual daily vehicle-meter counts within buffer meters of each s2 geohash.

Usage

get_traffic_summary(x, buffer = 400)

install_traffic()

Arguments

x

a vector of s2 cell identifers (s2_cell object)

buffer

distance from s2 cell (in meters) to summarize data

Value

a list the same length as x, which each element having a list of total_aadt_m and truck_aadt_m estimates

Details

A s2 level 14 approximation (~ 521 m sq) is used to simplify the intersection calculation with traffic summary data

References

https://www.fhwa.dot.gov/policyinformation/hpms.cfm

https://data-usdot.opendata.arcgis.com/datasets/usdot::highway-performance-monitoring-system-hpms-2020/about

https://www.fhwa.dot.gov/policyinformation/hpms/fieldmanual/hpms_field_manual_dec2016.pdf

Examples

get_traffic_summary(
  s2::as_s2_cell(c("8841b399ced97c47", "8841b38578834123")),
  buffer = 1500
)
#> $`8841b399ced97c47`
#> $`8841b399ced97c47`$aadt_total_m
#> [1] 657047536
#> 
#> $`8841b399ced97c47`$aadt_truck_m
#> [1] 4.305641e+14
#> 
#> 
#> $`8841b38578834123`
#> $`8841b38578834123`$aadt_total_m
#> [1] 225667410
#> 
#> $`8841b38578834123`$aadt_truck_m
#> [1] 2.010997e+14
#> 
#>