Get daily PM2.5 model predictions

Get daily PM2.5 model predictions

Usage

predict_pm25(x, dates)

Arguments

x

a vector of s2 cell identifers (s2_cell object)

dates

a list of date vectors for the predictions, must be the same length as x

Value

a list of tibbles the same length as x, each containing columns for the predicted (pm25) and its standard error (pm25_se); with one row per date in dates. These numerics are the concentrations of fine particulate matter, measured in micrograms per cubic meter. See vignette("cv-model-performance")

for more details on the cross validated accuracy of the daily PM2.5 model predictions.

Examples

d <- list(
  "8841b39a7c46e25f" = as.Date(c("2023-05-18", "2023-11-06")),
  "8841a45555555555" = as.Date(c("2023-06-22", "2023-08-15"))
)
predict_pm25(x = s2::as_s2_cell(names(d)), dates = d)
#> ℹ (down)loading random forest model
#> ✔ (down)loading random forest model [10.9s]
#> 
#> ℹ checking that s2 are within the contiguous US
#> ✔ checking that s2 are within the contiguous US [54ms]
#> 
#> ℹ adding coordinates
#> ✔ adding coordinates [10ms]
#> 
#> ℹ adding elevation
#> ✔ adding elevation [75ms]
#> 
#> ℹ adding HMS smoke data
#> ✔ adding HMS smoke data [959ms]
#> 
#> ℹ adding NARR
#> ✔ adding NARR [465ms]
#> 
#> ℹ adding gridMET
#> ✔ adding gridMET [433ms]
#> 
#> ℹ adding MERRA
#> ✔ adding MERRA [2.1s]
#> 
#> ℹ adding time components
#> ✔ adding time components [9ms]
#> 
#> [[1]]
#> # A tibble: 2 × 2
#>    pm25 pm25_se
#>   <dbl>   <dbl>
#> 1  7.76   0.837
#> 2  9.22   0.696
#> 
#> [[2]]
#> # A tibble: 2 × 2
#>    pm25 pm25_se
#>   <dbl>   <dbl>
#> 1  5.10   0.386
#> 2  5.75   0.843
#>